+++ StarLevelSystem/common/RandomSelect.cs
@@ -8,17 +8,17 @@ namespace StarLevelSystem.common
     internal static class RandomSelect
     {
         public static string RandomSelectFromWeightedListWithExclusions(List<ProbabilityEntry> listOfWeights, List<string> exclude) {
-            List<ProbabilityEntry> possibleModifiers = listOfWeights.Where(x => !exclude.Contains(x.Name)).ToList();
+            List<ProbabilityEntry> possibleModifiers = listOfWeights.Where(x => exclude.Contains(x.Name) == false).ToList();
             float totalweight = possibleModifiers.Select(x => x.SelectionWeight).Sum();
             if (totalweight == 0) { return CreatureModifiers.NoMods; }
             float selection = UnityEngine.Random.Range(0, totalweight);
             float current_weight = 0f;
-            // Logger.LogDebug($"Total weight is {totalweight}, random selection is {selection}");
+            //Logger.LogDebug($"Total weight is {totalweight}, random selection is {selection}");
             foreach (var entry in listOfWeights) {
                 current_weight += entry.SelectionWeight;
-                // Logger.LogDebug($"Current weight is {current_weight} >= {selection} for entry {entry.Name} - {entry.SelectionWeight}");
+                //Logger.LogDebug($"Current weight is {current_weight} >= {selection} for entry {entry.Name} - {entry.SelectionWeight}");
                 if (current_weight >= selection) {
-                    // Logger.LogDebug($"Randomly selected {entry.Name}");
+                    //Logger.LogDebug($"Randomly selected {entry.Name}");
                     return entry.Name;
                 }
             }